% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquarium_day.R
\name{aquarium_day}
\alias{aquarium_day}
\title{A function to report on your day at the aquarium!}
\usage{
aquarium_day(animals, time_of_day, amount)
}
\arguments{
\item{animals}{species of animal you saw}

\item{time_of_day}{time of day}

\item{amount}{number of that species you saw}
}
\value{
a sentence describing your time at the aquarium!
}
\description{
A function to report on your day at the aquarium!
}
\examples{
aquarium_day("octopi", "noon", 5)
}
